修改，用户中心模块（简历管理页面，用户设置，），上传文件管理

+++ bresume/bresume-core/src/main/java/com/bresume/core/common/base/dao/support/SimpleHibernateDao.java
@@ -335,11 +335,11 @@ public abstract class SimpleHibernateDao<T, ID extends Serializable> extends Abs
 	}
 
 	@Override
-	public int count()
+	public long count()
 	{
 		Criteria criteria = createCriteria();
 		criteria.setProjection(Projections.rowCount());
-		return (Integer) criteria.list().get(0);
+		return (Long) criteria.list().get(0);
 	}
 
 	@Override
